+++ trax/Grid.h
@@ -14,8 +14,8 @@ class Grid
 
   virtual double x(long i, long j) const = 0;
   virtual double y(long i, long j) const = 0;
-  virtual double operator()(long i, long j) const = 0;
-  virtual void set(long i, long j, double z) = 0;
+  virtual float operator()(long i, long j) const = 0;
+  virtual void set(long i, long j, float z) = 0;
   virtual bool valid(long i, long j) const = 0;
   virtual std::size_t width() const = 0;
   virtual std::size_t height() const = 0;
